ingestion_api_gcs/requirements.txt
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorama==0.4.6
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-core==2.4.1
google-cloud-secret-manager==2.20.2
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.2
grpcio-status==1.66.2
idna==3.10
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
proto-plus==1.24.0
protobuf==5.28.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pytest==8.3.3
requests==2.32.3
rsa==4.9
urllib3==2.2.3
